Some fixes done, some still left!
