+++ .github/scripts/add_footer.py
@@ -3,17 +3,22 @@ from os.path import join
 from utils import get_template
 from re import sub
 
-# The path to the vault. For now set to current working directory, as I assume all scripts are ran from the root path by default
-VAULT_PATH = getcwd()
 
-# Set to False to disable informational logging
-LOG = True
+from utils import get_root_of_vault
+# These are directories and files to exclude
+# By adding a dir/file, this script will ignore them and never change them!
+DIRECTORIES_TO_EXCLUDE = ['.git', '.github', '.idea', 'venv', 'DO NOT COMMIT']
+FILES_TO_EXCLUDE = ['.DS_Store', '.gitignore']
 
 
-def add_edit_on_github():
+def add_footer(root: str, debug: bool = True):
     """
-    Walks through the filetree rooted at VAULT_PATH.
+    Walks through the filetree rooted at `root`.
     For each markdown file that it finds, it replaces a particular comment line with the corresponding template.
+
+    Parameters:
+        root: path from which this method should run. Generally: root of the hub.
+        debug: boolean that indicates wether or not to print logging statements
     """
     # Grab the template
     template = get_template("footer")
@@ -22,7 +27,12 @@ def add_edit_on_github():
     comment = r"%% Hub footer: Please don't edit anything below this line %%"
 
     # Loop through the files
-    for root, _, files in walk(VAULT_PATH, topdown=True):
+    for root, dirs, files in walk(root, topdown=True):
+        # Exclude directories and files
+        dirs[:] = [d for d in dirs if d not in DIRECTORIES_TO_EXCLUDE]
+        files[:] = [f for f in files if f not in FILES_TO_EXCLUDE]
+
+        # Loop through the files
         for file in files:
             # We only care about markdown files
             # Note: Alternative implementation is to use os.splitext;
@@ -31,7 +41,7 @@ def add_edit_on_github():
                 # Get the full filepath
                 the_file = join(root, file)
 
-                if LOG:
+                if debug:
                     print(f"Processing '{the_file}'...")
 
                 # Get the rendered template
@@ -42,37 +52,35 @@ def add_edit_on_github():
                     # Read the file contents
                     contents = f.read()
 
-                    # Create variable holding the replacement text
-                    # By default, don't replace anything
-                    replacement = contents
-
-                    # If the comment is found
+                    # Check if our particular comment is present
                     if comment in contents:
-                        # And we don't have the entire template yet
-                        if render not in contents:
-                            replacement = sub(comment, render, contents)
-
-                            if LOG:
-                                print("\t=> Replacing the line with the template.")
-
+                        replacement = sub(comment, render, contents)
+                        if debug:
+                            print(
+                                f"\t=> Replacing everything below the line with the template for '{file}'.")
+                    # If it's not there: Add it
                     else:
-                        # Comment not found!
                         replacement = contents + "\n" + render
 
-                        if LOG:
-                            print("\t=> Adding the template.")
+                        if debug:
+                            print(f"\t=> Adding the template for '{file}'.")
 
                     # Actually write
                     # This is done by seeking to the beginning of the file
                     f.seek(0)
                     # Then writing the replacement string
                     f.write(replacement)
-                    # And finally truncating the file to close it, and remove possible dangling information (e.g. if replacement is shorter than original contents)
+                    # And finally truncating the file to close it
                     f.truncate()
 
 
 def main():
-    add_edit_on_github()
+    # Grab the root folder to run in.
+    # Uses the utility method to get root of the vault.
+    root = get_root_of_vault()
+    
+    # Set debug to whichever you'd like; true by default
+    add_footer(root)
 
 
 if __name__ == "__main__":
