+++ .github/scripts/templates/footer.md.jinja
@@ -1,7 +1,8 @@
 %% Hub footer: Please don't edit anything below this line %%
+##Footer
 
-<span class="github-edit-note">[Edit In Github](https://github.dev/obsidian-community/obsidian-hub/blob/{{ file_path }})</span>
+<span class="github-edit-note"> [Edit In GitHub](https://github.dev/obsidian-community/obsidian-hub/blob/{{ file_path }})</span>
 
 <span class="github-copy-note"> [Copy this note](https://raw.githubusercontent.com/obsidian-community/obsidian-hub/main/{{ file_path }})</span>
 
-<span class="github-vault-download">[Download this vault](https://github.com/obsidian-community/obsidian-hub/archive/refs/heads/main.zip)</span>
+<span class="github-vault-download"> [Download this vault](https://github.com/obsidian-community/obsidian-hub/archive/refs/heads/main.zip)</span>
